IndexedDB/tests/submissions/Opera/idbobjectstore_createIndex3.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBObjectStore.createIndex() - the index is usable right after being made</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>
    var db, aborted,
      t = async_test()

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        var txn = e.target.transaction,
          objStore = db.createObjectStore("store", { keyPath: 'key' });

        for (var i = 0; i < 1000; i++)
            objStore.add({ key: "key_" + i, indexedProperty: "indexed_" + i });

        objStore.createIndex("index", "indexedProperty")
                .get('indexed_999')
                .onsuccess = t.step_func(function(e) {
            assert_equals(e.target.result.key, 'key_999', 'key');
        });
    }

    open_rq.onsuccess = function() {
        t.done();
    }
</script>

<div id="log"></div>
